+++ server/lib/handlebars-helpers/index.js
@@ -1,8 +1,8 @@
 module.exports = {
 
+	instanceNamingValue: require('./instance-naming-value'),
 	model: require('./model'),
 	modelLowerCase: require('./model-lower-case'),
-	modelNamingValue: require('./model-naming-value'),
 	modelRoute: require('./model-route')
 
 };
