+++ server/lib/handlebars-helpers/instance-naming-value.js
@@ -1,0 +1,9 @@
+const modelNamingPropMap = require('../model-naming-prop-map');
+
+module.exports = function (instance) {
+
+	const model = instance.constructor.name.toLowerCase();
+
+	return instance[modelNamingPropMap[model]];
+
+};
