--- server/lib/handlebars-helpers/model-naming-value.js
@@ -1,9 +1,0 @@
-const pageTitleTextMap = require('../page-title-text-map');
-
-module.exports = function (instance) {
-
-	const model = instance.constructor.name.toLowerCase();
-
-	return instance[pageTitleTextMap[model]];
-
-};
