+++ views/partials/instance-link.html
@@ -1,1 +1,1 @@
-<a href="/{{modelRoute instance}}/{{instance.id}}">{{modelNamingValue instance}}</a>
+<a href="/{{modelRoute instance}}/{{instance.id}}">{{instanceNamingValue instance}}</a>
